+++ design-system/logo/src/atlassian-analytics-logo/index.tsx
@@ -1,0 +1,2 @@
+export { AtlassianAnalyticsIcon } from './icon';
+export { AtlassianAnalyticsLogo } from './logo';