+++ design-system/logo/src/jira-logo/icon.tsx
@@ -27,7 +27,6 @@ const svg = ({
 
   if (appearance) {
     colors = getColorsFromAppearance(appearance);
-    id = `jiraIcon-${appearance}`;
   }
 
   return `