+++ design-system/logo/src/jira-product-discovery-logo/logo.tsx
@@ -28,7 +28,6 @@ const svg = ({
 
   if (appearance) {
     colors = getColorsFromAppearance(appearance);
-    id = `jpdLogo-${appearance}`;
   }
 
   return `