+++ design-system/logo/src/opsgenie-logo/new-logo.tsx
@@ -29,7 +29,6 @@ const svg = ({
 
   if (appearance) {
     colors = getColorsFromAppearance(appearance);
-    id = `opsgenieLogo-${appearance}`;
   }
 
   return `