+++ linking-platform/link-client-extension/CHANGELOG.md
@@ -1,1 +1,11 @@
 # @atlaskit/link-client-extension
+
+## 0.2.0
+
+### Minor Changes
+
+- [`f32cef73d0f`](https://bitbucket.org/atlassian/atlassian-frontend/commits/f32cef73d0f) - use new types from smart link types
+
+### Patch Changes
+
+- Updated dependencies